Updated dependencies for Django 3
Additionally, updated many other dependencies in order to
support Django 3. This required me to also update
pysidian-core to version 0.0.10.

Moreover, I manually changed the hash in the Pipfile.lock for
the pysidian-core version 0.0.10 package, because it was
(for some reason) using an md5 hash, which has been disabled
in a recent pip version. This bug has been fixed in pipenv,
but appears to not have been shiped to pypi yet. As such,
updating or locking would currently cause a breakage.